refactor api entry function names
