+++ tests/aa_pbs_exporter/pbs_2022_01/parse/pages/three_pages/three_page_test.py
@@ -6,7 +6,7 @@ from pydantic import parse_raw_as
 from tests.aa_pbs_exporter.resources.helpers_2 import ResourceTestData
 
 from aa_pbs_exporter.pbs_2022_01.models import raw
-from aa_pbs_exporter.pbs_2022_01.parse_pbs_txt import parse_pbs_txt_file
+from aa_pbs_exporter.pbs_2022_01.parse_pbs_txt_file import parse_pbs_txt_file
 
 SERIALIZE_ONLY = False
 TEST_DATA = ResourceTestData("three_pages.txt", "02MAY2022_MIA_raw.json")
